+++ StarkBank/StarkBank/BoletoHolmes/BoletoHolmes.cs
@@ -193,8 +193,8 @@ namespace StarkBank
                 resourceMaker: resourceMaker,
                 query: new Dictionary<string, object> {
                     { "limit", limit },
-                    { "after", after },
-                    { "before", before },
+                    { "after", new Utils.StarkBankDate(after) },
+                    { "before", new Utils.StarkBankDate(before) },
                     { "tags", tags },
                     { "ids", ids},
                     { "status", status },