+++ StarkBank/StarkBank/Deposit/Log.cs
@@ -116,8 +116,8 @@ namespace StarkBank
                     resourceMaker: resourceMaker,
                     query: new Dictionary<string, object> {
                         { "limit", limit },
-                        { "after", after },
-                        { "before", before },
+                        { "after", new Utils.StarkBankDate(after) },
+                        { "before", new Utils.StarkBankDate(before) },
                         { "types", types },
                         { "depositIds", depositIds }
                     },