+++ StarkBank/StarkBank/Event/Event.cs
@@ -113,8 +113,8 @@ namespace StarkBank
                 resourceMaker: resourceMaker,
                 query: new Dictionary<string, object> {
                     { "limit", limit },
-                    { "after", after },
-                    { "before", before },
+                    { "after", new Utils.StarkBankDate(after) },
+                    { "before", new Utils.StarkBankDate(before) },
                     { "isDelivered", isDelivered }
                 },
                 user: user